# add zstd compression codec