@xiaokang 2022-05-24
https://github.com/apache/incubator-doris/pull/8923 need to be merged before users can enable zstd or other codec, since current Lz4f codec is hardcoded and not configurable.